Fix bug in splitting of internal nodes.
When having to split an internal node due to an insertion into this same
node, then the newly created separator key (of the split internal node)
will dictate which of the two fresh nodes the to-be-inserted key must be
inserted.

The existing implementation chose the insertion target based on the
range of values, which did not necessarily align with the new separator
key.

This commit also changes the test-root-node-split test into a benchmark,
as the volume of insertions required for a meanignful amount (read: two)
of root splits to happen will take a few dozen seconds.